+++ src/workspaceSearcher.ts
@@ -1,6 +1,7 @@
-import { Uri, workspace } from "vscode";
+import { CancellationTokenSource, CancellationToken, Uri, workspace } from "vscode";
 import { TypeHintSettings } from "./settings";
 import { TypeSearch } from "./typeSearch";
+import { DataTypeContainer } from "./python";
 
 /**
  * Searches Python files in the workspace.
@@ -9,17 +10,15 @@ export class WorkspaceSearcher {
 
     private search: boolean = true;
     private activeDocUri: Uri;
+    private typeContainer: DataTypeContainer;
+    private tokenSource: CancellationTokenSource;
     private settings: TypeHintSettings;
 
-    /**
-     * Constructs a new WorkspaceSearcher.
-     * 
-     * @param activeDocumentUri The uri of the active document.
-     * @param settings User settings.
-     */
-    constructor(activeDocumentUri: Uri, settings: TypeHintSettings) {
+    constructor(activeDocumentUri: Uri, settings: TypeHintSettings, typeContainer: DataTypeContainer) {
         this.activeDocUri = activeDocumentUri;
         this.settings = settings;
+        this.tokenSource = new CancellationTokenSource();
+        this.typeContainer = typeContainer;
     }
 
     /**
@@ -31,18 +30,22 @@ export class WorkspaceSearcher {
      */
     public async findHintOfSimilarParam(param: string, activeDocumentText: string): Promise<string | null> {
         this.search = true;
-        const maxResults = this.settings.fileSearchLimit;
+        const maxResults = this.settings.workspaceSearchLimit;
 
         if (maxResults > 0 && workspace.workspaceFolders) {
-            const uris = (await workspace.findFiles("**/*.py", null, maxResults))
-                .filter(u => u.path !== this.activeDocUri.path);
+            const uriSplit = this.activeDocUri.path.split("/");
+            const glob = `**/${uriSplit[uriSplit.length - 1]}`;
+            const uris = await workspace.findFiles("**/*.py", glob, maxResults, this.tokenSource.token);
 
             for (let i = 0; this.search && i < uris.length; i++) {
                 let doc = await workspace.openTextDocument(uris[i]);
                 let docText = doc.getText();
                 let type = TypeSearch.hintOfSimilarParam(param, docText);
-                if (type) {
-                    if (type = TypeSearch.findImport(type, activeDocumentText, false)) {
+                if (this.search && type) {
+                    if (!(type in this.typeContainer)) {
+                        type = TypeSearch.findImport(type, activeDocumentText, false);
+                    }
+                    if (type) {
                         return type;
                     }
                 }
@@ -55,6 +58,7 @@ export class WorkspaceSearcher {
      * Stops all searches.
      */
     public cancel() {
+        this.tokenSource.cancel();
         this.search = false;
     }
 }