+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [1.1.0] - 2020-04-19
+
+### Improvements:
+
+* Added a new setting for disabling workspace searching.
+* More default hints are provided for the typing module.
+* Overall speed improvements.
+
+### Fixes:
+
+* Setting the workspace search limit to 0 not working.
+
 ## [1.0.2] - 2020-04-16
 
 * Fixed type hints being provided for non-parameters.